First contribution
As we talked on the call there are several changes.
The visuals are easy to change and is all implemented on the same scene.
Tags added: Enemy and Powerup
Layers added: Environment, Player, Pickups
The Layer Environment is for the colliders that limit the player movement area. On physics preferences, Enemies don't collide with Environment or Pickups